requirements.txt
scrapy==2.5.1
requests~=2.27.1
Django==4.0.1
djangorestframework
markdown
django-filter
